requirements.txt
robotframework==7.1.1
robotframework-browser==18.9.1
robotframework-jsonlibrary==0.5
Faker==30.6.0
python_dotenv==1.0.1




